Gap 12 phase D: memory64 support
memory64 modules ((memory i64 N)) now run end-to-end through the
interpreter and transpiler when paired with NativePointer storage.
The cap shifts from WasmMaxPages (2^16, 4 GiB for memory32) to
WasmMaxPages64 (2^48). All 4 spec.test fixtures under
spec/test/core/memory64/ pass on both WAST and WAST-transpiled
paths.

Bounds checks across every memory-op site (single-byte, narrow,
full-width, SIMD, atomics, bulk init/copy/fill) now use a wrap-
safe unsigned form:
    (ulong)ea > (ulong)mem.ByteLength
    || (ulong)mem.ByteLength - (ulong)ea < (ulong)width
which traps for both negative ea (cast to huge ulong) and
near-2^64 ea where ea+width would wrap.

OpStack.PopAddr no longer traps on negative — memory64 addresses
with the high bit set are valid wasm. Bounds-check sites are now
responsible for trapping on out-of-bounds.

MemoryInstance.MaxPagesForMode reads Type.Limits.AddressType and
returns WasmMaxPages64 (2^48) for i64-address NativePointer
memories; WasmMaxPages (2^16) for i32; HostMaxPages (2^15) for
ManagedArray either way (byte[] cap is hard).

InstTableGet / InstTableSet switched to unsigned compare so
table64 ((table i64 …)) indices behave correctly. Spec
table_get64.wast and table_set64.wast now pass.

Three new tests in MemoryNativePointerEndToEndTests cover
memory64 page cap, high-address store/load round-trip, and
out-of-bounds trap. Wacs.Core 394 (was 391), Spec.Test 770/772,
every other suite unchanged.

WACS 0.12.3 → 0.12.4.

Phase D.2 (out of scope for this commit): widen transpiler-emit
memory-op helper sigs from int addr, long offset to nuint/ulong
for memory64 modules with addresses past int.MaxValue;
WacsHostMemory.AsSpan int-bounded surface gets a >2 GiB
MemoryHandle companion.

+++ Wacs.Core/Wacs.Core/Instructions/Table.cs
@@ -64,8 +64,13 @@ namespace Wacs.Core.Instructions
                 $"Instruction table.get failed. Wrong type on stack.");
             //7.
             long i = context.OpStack.PopAddr();
-            //8.
-            if (i >= tab.Elements.Count || (i > (long)int.MaxValue))
+            //8. Phase D: unsigned compare. table64 indices ride on
+            // the same OpStack pop path; PopAddr no longer traps on
+            // negative (memory64-compatible). For tables, treat i
+            // as u64 — negative-as-signed maps to a huge ulong that
+            // always fails the bound.
+            if ((ulong)i >= (ulong)tab.Elements.Count
+                || i > (long)int.MaxValue)
             {
                 throw new TrapException("Trap in table.get");
             }
@@ -130,8 +135,8 @@ namespace Wacs.Core.Instructions
                 $"Instruction table.set found incorrect type on top of the Stack");
             //9.
             long i = context.OpStack.PopAddr();
-            //10.
-            if (i >= tab.Elements.Count)
+            //10. Phase D: unsigned compare for table64 indices.
+            if ((ulong)i >= (ulong)tab.Elements.Count)
             {
                 throw new TrapException("table.set index exceeds table size.");
             }
